@@ -4,10 +4,10 @@ namespace Application.DTOs
 {
     public class PigIntakeInsertDTO
     {
-        [Required(ErrorMessage = " ExpectedQuantity is required")]
+        [Required(ErrorMessage = "Số lượng dự kiến bắt buộc")]
         public int ExpectedQuantity { get; set; }
 
-        [Required(ErrorMessage = " PigName is required")]
+        [Required(ErrorMessage = "Mã khu vực bắt buộc")]
         public string AreasId { get; set; }
 
     }
