@@ -8,10 +8,10 @@ namespace Application.Interfaces
     {
         Task<BasePagination<PigInTakeModelView>> GetAllAsync(int pageIndex, int pageSize, string filter);
         Task<PigInTakeModelView> GetPigIntakeByIdAsync(string id);
-        Task InsertIntakeAsync(PigIntakeInsertDTO dTO);
+        Task<PigInTakeModelView> InsertIntakeAsync(PigIntakeInsertDTO dTO);
         Task<PigDeliveryModel> UpdateIntakeAsync(string id, PigIntakeUpdateDTO model);
-        Task AllocatePigsToStableAsync(string AreasId, string pigIntakeId);
-        Task AcceptIntakeAsync(string id, PigIntakeAcceptDTO model);
+        Task<PigInTakeModelView> AllocatePigsToStableAsync(string AreasId, string pigIntakeId);
+        Task<PigInTakeModelView> AcceptIntakeAsync(string id, PigIntakeAcceptDTO model);
         Task DeleteAsync(string id);
     }
 }
