@@ -1,18 +1,45 @@
+using System.Text.Json.Serialization;
+
 namespace Application.Models
 {
     public class PigInTakeModelView
     {
         public string Id { get; set; }
-        public string SuppliersId { get; set; }
+        [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
+
+        public string SuppliersName { get; set; }
+        [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingDefault)]
 
         public decimal UnitPrice { get; set; }
+        [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingDefault)]
+
         public decimal Deposit { get; set; }
+        [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingDefault)]
+
         public decimal TotalPrice { get; set; }
+        [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingDefault)]
+
 
         public int ExpectedQuantity { get; set; }
+        [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingDefault)]
+
         public int ReceivedQuantity { get; set; }
+        [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingDefault)]
+
         public int AcceptedQuantity { get; set; }
+        [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingDefault)]
+
         public int RejectedQuantity { get; set; }
+        [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingDefault)]
+
+
         public DateTimeOffset ApprovedTime { get; set; }
+        [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingDefault)]
+
+        public DateTimeOffset DeliveryDate { get; set; }
+        [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingDefault)]
+
+        public DateTimeOffset StokeDate { get; set; }
+
     }
 }
